Prevent file loss/exceptions when renaming: (#278)
Do nothing if renaming to the same path
Explicit exception when attempting to rename a dir

+++ cloudpathlib/cloudpath.py
@@ -453,6 +453,15 @@ class CloudPath(metaclass=CloudPathMeta):
                 f"The target based to rename must be an instantiated class of type: {type(self)}"
             )
 
+        if self.is_dir():
+            raise CloudPathIsADirectoryError(
+                f"Path {self} is a directory; rename/replace the files recursively."
+            )
+
+        if target == self:
+            # Request is to replace/rename this with the same path - nothing to do
+            return self
+
         if target.exists():
             target.unlink()
